% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRTemplate.R
\name{makeRTemplate}
\alias{makeRTemplate}
\title{Create a .R template}
\usage{
makeRTemplate(
  r.dir = getwd(),
  r.name = NULL,
  proj.name = NULL,
  start.date = format(Sys.Date(), "\%B \%d, \%Y"),
  version = "1",
  client = NULL,
  client.dept = NULL,
  main.statistician = NULL,
  stats.collab = NULL,
  r.purpose = NULL,
  r.notes = NULL
)
}
\arguments{
\item{r.dir}{A character string specifying the file path where the template
will be stored. Defaults to {getwd()}.}

\item{r.name}{NULL or a character string specifying the name of the .R file.
If NULL, the file will be named p[proj.name]_r[currentdate]_v[version] or
r[currentdate].}

\item{proj.name}{NULL or a character string providing the name of the project
to be included in the template name and/or header.}

\item{start.date}{NULL or a character string providing the date to be included
in the template header. Defaults to today's date.}

\item{version}{NULL or a character string providing the project version to be
included in the template name and/or header. Defaults to "1".}

\item{client}{NULL or a character string providing the name(s) of client(s) to
be included in the template header.}

\item{client.dept}{NULL or a character string providing the affiliation of the
client(s) to be included in the template header.}

\item{main.statistician}{NULL or a character string providing the name of the
primary statistician. This will be used in the template header.}

\item{stats.collab}{NULL or a character string providing additional 
collaborators to be included as authors in the template header.}

\item{r.purpose}{NULL or a character string providing the purpose for the
generated template to be included in the template header.}

\item{r.notes}{NULL or a character string providing additional notes to be
included in the template header.}
}
\description{
`makeRTemplate` will generate a .R file with a header containing
   any project information provided in the function call.
}
\details{
You can append code and comments under the generated header by either
   creating a .basefile.R file in your /home/<USERNAME> directory or by modifying
    the templates/basefile.R file contained in the package library.
    
    To find the location of your HOME directory, run
    `Sys.getenv("HOME")`.

    To find the location of the included template file, run
    `system.file("templates", package = "startProject")`.
}
\examples{
\dontrun{makeRTemplate(r.name = "rTemplate")}

}
\author{
Rocio Lopez, \email{Rocio.LopezMoscoso@cuanschutz.edu}
}
