% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp_startProject.R
\name{runApp_startProject}
\alias{runApp_startProject}
\title{Launch startProject Shiny App}
\usage{
runApp_startProject(browser = "WindowViewer")
}
\arguments{
\item{browser}{A character string that specifies the browser in which the app
is launched. Valid options are "PaneViewer" to run in RStudio Viewer tab,
"WindowViewer" to run in RStudio Window or "WindowExternal" to run in
default web browser (case is ignored).
Defaults to "WindowViewer".}
}
\description{
`runApp_startProject` will open a viewer window that will allow you
    to run \code{\link{startProject}} as a local shiny App.
}
\details{
The browser argument only takes effect if running within RStudio.
}
\examples{
\dontrun{runApp_startProject(browser = "PaneViewer")}

}
\seealso{
\code{\link{startProject}} for details on inputs.
}
\author{
Rocio Lopez, \email{lopezr@ccf.org}
}
