% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startProject-package.R
\docType{package}
\name{startProject-package}
\alias{startProject-package}
\title{startProject: A package for creating a project directory}
\description{
The start.project package can be used to create a project directory and
generate memo and code templates.
}
\section{startProject Package Functions}{


    Functions to create a project directory with sub-folders and templates:

    \code{\link{startProject}}: A function to create project directory.

    \code{\link{runApp_startProject}}: A function to launch \code{\link{startProject}}
    as a local Shiny app.

    Functions to create additional templates:

    \code{\link{makeMemoTemplate}}: A function to generate a .docx memo template.

    \code{\link{makeRmdTemplate}}: A function to generate a .Rmd template.

    \code{\link{makeRTemplate}}: A function to generate a .R template.

    \code{\link{makeSasTemplate}}: A function to generate a .sas template.
}

\section{Details}{

    Will link web page/tutorial here.
}

\author{
Rocio Lopez, \email{Rocio.LopezMoscoso@cuanschutz.edu}
}
\keyword{internal}
